+++ src/main/java/br/com/meli/dhprojetointegrador/repository/BatchStockRepository.java
@@ -1,9 +1,7 @@
 package br.com.meli.dhprojetointegrador.repository;
 
+import br.com.meli.dhprojetointegrador.entity.BatchStock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.com.meli.dhprojetointegrador.entity.BatchStock;
-
-public interface BatchStockRepository extends JpaRepository<BatchStock, Integer> {
-
+public interface BatchStockRepository extends JpaRepository<BatchStock, Long> {
 }
